AutoModo-App/efi-django/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
